Extracts InvalidRecipe exception
